+++ src/components/dictionaries/liangan/engine.ts
@@ -8,7 +8,7 @@ export default function search (
   text: string,
   config: AppConfig
 ): Promise<DictSearchResult<LiangAnResult>> {
-  return moedictSearch<LiangAnResult>('c', encodeURIComponent(text), config)
+  return moedictSearch<LiangAnResult>('c', encodeURIComponent(text.replace(/\s+/g, '')), config)
     .then(result => {
       if (result.result.h) {
         result.result.h.forEach(h => {